Move VerifyEx from test file
